orbitlex-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
sgp4==2.22
numpy==1.26.4
scipy==1.13.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
groq==0.6.0
reportlab==4.2.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.7.0
